# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservation": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Get user reservations",
                "operationId": "28750d1ef9043e60393d17cae7bfaa23",
                "responses": {
                    "200": {
                        "description": "User reservations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 13
                                            },
                                            "space_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "event_name": {
                                                "type": "string",
                                                "example": "Some festival"
                                            },
                                            "start_date": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-10-02 12:30:00"
                                            },
                                            "end_date": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-10-02 15:00:00"
                                            },
                                            "space_name": {
                                                "type": "string",
                                                "example": "Big saloon"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Create a reservation",
                "operationId": "1609a7fc7630f503e45c5f6d2d4cde08",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "space_id",
                                    "event_name",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "space_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "event_name": {
                                        "type": "string",
                                        "example": "Some festival"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-10-02 12:30:00"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-10-02 15:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "result message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservation/{id}": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Get particular user reservation",
                "operationId": "b14230469fb85e8a018c7c5f903dea80",
                "responses": {
                    "200": {
                        "description": "User reservation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 13
                                        },
                                        "space_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "event_name": {
                                            "type": "string",
                                            "example": "Some festival"
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-02 12:30:00"
                                        },
                                        "end_date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-02 15:00:00"
                                        },
                                        "space_name": {
                                            "type": "string",
                                            "example": "Big saloon"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Update a reservation",
                "operationId": "59ed340248fdebd6e9c2ef7b3a8e2f23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "space_id",
                                    "event_name",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "space_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "event_name": {
                                        "type": "string",
                                        "example": "Some festival"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-10-02 12:30:00"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-10-02 15:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Result message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Delete reservation",
                "operationId": "ef1c18f96166fc3677496981d3cd4542",
                "responses": {
                    "200": {
                        "description": "Result messge",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}